backend/requirements.txt
fastapi[all]
requests
pytest
httpx
psycopg2-binary
boto3
python-jose[cryptography]
passlib[bcrypt]
python-multipart
matplotlib
plotly
mplcursors
mpld3